FBP-11. Implemented basic resizing (no cascade)
