+++ bpmn-intellij-plugin/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/render/elements/anchors/ShapeResizeAnchorBottom.kt
@@ -33,7 +33,7 @@ class ShapeResizeAnchorBottom(
             return mutableMapOf()
         }
 
-        return super.render(ctx)
+        return super.doRender(ctx)
     }
 
     override fun icon(): Icon {
