+++ bpmn-intellij-plugin/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/render/elements/anchors/ShapeResizeAnchorTop.kt
@@ -37,7 +37,7 @@ class ShapeResizeAnchorTop(
             return mutableMapOf()
         }
 
-        return super.render(ctx)
+        return super.doRender(ctx)
     }
 
     override fun icon(): Icon {
